+++ mapie/tests/test_utils.py
@@ -7,17 +7,17 @@ import pytest
 from numpy.random import RandomState
 from sklearn.datasets import make_regression
 from sklearn.linear_model import LinearRegression
+from sklearn.model_selection import BaseCrossValidator
 from sklearn.utils.validation import check_is_fitted
 
 from mapie._typing import ArrayLike, NDArray
 from mapie.regression import MapieQuantileRegressor
 from mapie.utils import (check_alpha, check_alpha_and_n_samples,
-                         check_binary_zero_one, check_lower_upper_bounds,
-                         check_n_features_in, check_n_jobs, check_null_weight,
-                         check_number_bins, check_split_strategy,
-                         check_verbose, compute_quantiles, fit_estimator,
-                         get_binning_groups)
-
+                         check_binary_zero_one, check_cv,
+                         check_lower_upper_bounds, check_n_features_in,
+                         check_n_jobs, check_null_weight, check_number_bins,
+                         check_split_strategy, check_verbose,
+                         compute_quantiles, fit_estimator, get_binning_groups)
 
 X_toy = np.array([0, 1, 2, 3, 4, 5]).reshape(-1, 1)
 y_toy = np.array([5, 7, 9, 11, 13, 15])
@@ -32,8 +32,8 @@ ALPHAS = [
     np.array([0.05, 0.1, 0.2]),
 ]
 
-
-prng = RandomState(1234567890)
+random_state = 1234567890
+prng = RandomState(random_state)
 y_score = prng.random(51)
 y_scores = prng.random((51, 5))
 y_true = prng.randint(0, 2, 51)
@@ -410,3 +410,33 @@ def test_change_values_zero_one() -> None:
     """Test that binary output are changed to zero one outputs."""
     array_ = check_binary_zero_one(np.array([0, 4, 4]))
     assert (np.unique(array_) == np.array([0, 1])).all()
+
+
+@pytest.mark.parametrize("cv", [5, "split"])
+def test_check_cv_same_split_with_random_state(cv: BaseCrossValidator) -> None:
+    """Test that cv generate same split with fixed random_state."""
+    cv = check_cv(cv, random_state=random_state)
+
+    train_indices_1, train_indices_2 = [], []
+    for train_index, _ in cv.split(X):
+        train_indices_1.append(train_index)
+    for train_index, _ in cv.split(X):
+        train_indices_2.append(train_index)
+
+    for i in range(cv.get_n_splits()):
+        np.testing.assert_allclose(train_indices_1[i], train_indices_2[i])
+
+
+@pytest.mark.parametrize("cv", [5, "split"])
+def test_check_cv_same_split_no_random_state(cv: BaseCrossValidator) -> None:
+    """Test that cv generate same split with no random_state."""
+    cv = check_cv(cv, random_state=None)
+
+    train_indices_1, train_indices_2 = [], []
+    for train_index, _ in cv.split(X):
+        train_indices_1.append(train_index)
+    for train_index, _ in cv.split(X):
+        train_indices_2.append(train_index)
+
+    for i in range(cv.get_n_splits()):
+        np.testing.assert_allclose(train_indices_1[i], train_indices_2[i])