+++ HISTORY.rst
@@ -11,6 +11,7 @@ History
 * Refactor MapieRegressor and ConformityScore to add the possibility to use X in ConformityScore.
 * Separate the handling of the estimator from MapieRegressor into a new class called EnsembleEstimator.
 * Fix an unfixed random state in one of the classification tests.
+* Fix and preserve the split behavior of the check_cv method with and without random state.
 
 0.6.5 (2023-06-06)
 ------------------